+++ src/lib/zsh/commands.ts
@@ -46,6 +46,13 @@ export const getCommands = async (): Promise<
   )
   const { certsImportCommand } = await import('../../commands/certs/import.ts')
   const { certsRmCommand } = await import('../../commands/certs/rm.ts')
+  const { certsCaInstallCommand } = await import(
+    '../../commands/certs/ca/install.ts'
+  )
+  const { certsCaUninstallCommand } = await import(
+    '../../commands/certs/ca/uninstall.ts'
+  )
+  const { certsCaInfoCommand } = await import('../../commands/certs/ca/info.ts')
 
   return {
     run: runCommand,
@@ -73,5 +80,8 @@ export const getCommands = async (): Promise<
     'certs:generate': certsGenerateCommand,
     'certs:import': certsImportCommand,
     'certs:rm': certsRmCommand,
+    'certs:ca:install': certsCaInstallCommand,
+    'certs:ca:uninstall': certsCaUninstallCommand,
+    'certs:ca:info': certsCaInfoCommand,
   }
 }